requirements.txt
aiogram
tortoise-orm